rename: RoomDetailTab 폴더 생성
