+++ frontend/src/domains/room/components/RoomDetailTab/ProgressPickeat.tsx
