+++ frontend/src/domains/room/components/RoomDetailTab/index.tsx
